re-naming PaginatedQueryInterface to PageableQuery
